Made PSR-0 loading more standards-conform
This solves deprecation notices in Composer 1.10+ and allows languages
to still be loaded using Composer's autoloader when Composer 2 is
released.

It changes the language verification by checking HTMLPurifier/Languages
as well as HTMLPurifier/Languages/classes/, where the latter was the
previous storage location. Both structures are supported for backwards
compatibility.

Fixes #246